Introducing a new One-Shot Patch Method.
Upon importation of vulnerability results (from Nessus or Nexpose) and successful exploitation of a single vulnerability, 
you can build an entire patching script, which will patch all current vulnerabilities.

This has become the "default" run style, if however you'd like to run in the previous mode (now refereed to as "Solo") you may do so by loading unsploitable_solo (or using the solo launch script).

Commands:
        - unsploit_patch        => Build Missing Patches (based on vulnerabilities).
        - unsploit_show         => Show Matched Patches (based on vulnerabilities).

After building the patch script you can upload and execute the script.  Curl is a prerequisite on Windows systems (and it is included in the .curl directory) - so don't forget to upload Curl to the same directory before launching the patch script.

Additionally patches.db has been updated to support the new functionality as well as correct a few issues with Windows 2000 and Windows 2003